Move search functions to search module
